Refactor random number generation

+++ tests/compile-fail/getrandom.rs
@@ -4,7 +4,7 @@ extern crate libc;
 fn main() {
     let mut buf = [0u8; 5];
     unsafe {
-        libc::syscall(libc::SYS_getrandom, buf.as_mut_ptr() as *mut libc::c_void, 5, 0);
+        libc::syscall(libc::SYS_getrandom, buf.as_mut_ptr() as *mut libc::c_void, 5 as libc::size_t, 0 as libc::c_uint);
         //~^ ERROR constant evaluation error: miri does not support random number generators in deterministic mode!
     }
 }
